Fixes for MSVC 2013: All examples not depending from third-party libraries (boost, opencv) compile. Still many warnings related to int and long size on win64 platform.

+++ ff/gt.hpp
@@ -456,7 +456,7 @@ public:
         svector<ff_node*> _workers(nw);
         for(ssize_t i=0;i<running;++i) 
             if (!offline[i]) _workers.push_back(workers[i]);
-        svector<int> retry(nw);
+        svector<size_t> retry(nw);
 
         for(register size_t i=0;i<nw;++i) {
             if(i!=(size_t)channelid && !_workers[i]->get(&V[i]))
